+++ .eslintignore
@@ -2,4 +2,5 @@ dist/
 lib/
 node_modules/
 jest.config.js
-src/generator/parser/grammar.ts
+src/generator/parser/expression-parser.ts
+src/generator/parser/template-parser.ts
